refactor: refactor screens, extract styles
