+++ src/navigation/NavigationComponent.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import Home from "../screens/Home";
-import Favorites from "../screens/Favorites";
+import Home from "../screens/Home/Home";
+import Favorites from "../screens/Favorites/Favorites";
 import NavigationButton from "../components/NavigationButton/NavigationButton";
 
 const Stack = createNativeStackNavigator();
