+++ src/screens/Favorites/Favorites.style.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    MainContainer: {
+        paddingTop: 10,
+        backgroundColor: 'white'
+    },
+    Text: {
+        fontSize: 18,
+        textAlign: 'center',
+        marginTop: 50,
+    }
+})
