+++ src/screens/Home/Home.style.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+
+export const styles = StyleSheet.create({
+    MainContainer: {
+        height: '100%',
+        backgroundColor: 'white',
+        alignItems: 'center'
+    },
+    Text: {
+        fontSize: 18,
+        textAlign: 'center',
+        marginTop: 40,
+    }
+})
